masterinterface/requirements.txt
django>=1.3.1
oauth2>=1.5.167
python-openid>=2.2
django-social-auth

